Added entry points and plugin stuff.
