+++ annodize/dataframe/example/plugin.py
@@ -1,0 +1,12 @@
+"""Example `annodize.dataframe` plugin implementation.
+
+Note that this is specified in the `entry_points` of your package.
+"""
+
+
+def col_major():
+    """Column-major plugin entry point."""
+
+
+def row_major():
+    """Row-major plugin entry point."""
