+++ annodize/dataframe/plugin_loader.py
@@ -1,0 +1,1 @@
+"""Loads `annodize.dataframe` plugins."""
