+++ entry_points.ini
@@ -1,0 +1,5 @@
+# Plugin definition file
+
+[annodize.frame]
+example_col_major = annodize.dataframe.example.plugin:col_major
+example_row_major = annodize.dataframe.example.plugin:row_major
